couille makefile et memcmp

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: gnyssens <gnyssens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 12:54:11 by gnyssens          #+#    #+#             */
-/*   Updated: 2024/04/15 18:49:55 by gnyssens         ###   ########.fr       */
+/*   Updated: 2024/04/19 17:35:39 by gnyssens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_memcmp(const void *str1, const void *str2, size_t n)
 {
 	unsigned char	*s1;
 	unsigned char	*s2;
-	int				i;
+	size_t			i;
 
 	if ((!str1 && !str2) || n == 0)
 		return (0);
